refactored POM login
