+++ tests/registration.spec.ts
@@ -1,5 +1,7 @@
 import {test, expect} from '@playwright/test'
-import { faker } from '@faker-js/faker'
+import {faker} from '@faker-js/faker'
+import { RegistrationPage } from '../pageModels/registrationPage'
+
 
 test.beforeEach(async({page})=>{
     if (await page.getByRole('button', { name: "Поділитися" }).isVisible()) {
@@ -11,81 +13,60 @@ test.beforeEach(async({page})=>{
     }
 })
 
+
 test('valid registration', async ({page})=>{
-    const validUserName = faker.person.firstName()
-    const validUserPassword = faker.internet.password()
-    await page.getByPlaceholder('Ім\'я користувача').fill(validUserName);
-    await page.getByPlaceholder('Пароль').nth(0).fill(validUserPassword);
-    await page.getByPlaceholder('Повторіть пароль').fill(validUserPassword);
-    await page.getByRole('button', {name:"Зареєструватись"}).click()
+    const registrationPage = new RegistrationPage(page);
+    const validUserName = faker.person.firstName();
+    const validUserPassword = faker.internet.password();
 
-    await expect(page, "Redirected to main page after the login").toHaveURL('http://localhost:8000/')
-    await page.locator("#menu-button").click()
-    await expect(page.getByText(validUserName), "Name of the logged in user is visible").toBeVisible()
+    await registrationPage.register(validUserName,validUserPassword,validUserPassword);
+
+    await expect(page, "Redirected to main page after the login").toHaveURL('http://localhost:8000/');
+    await page.locator("#menu-button").click();
+    await expect(page.getByText(validUserName), "Name of the logged in user is visible").toBeVisible();
 })
 
-test('not valid password registration(short)', async ({page})=>{
-    const validUserName = faker.person.firstName()
-    const shortUserPassword = 'par0l'
-    await page.getByPlaceholder('Ім\'я користувача').fill(validUserName);
-    await page.getByPlaceholder('Пароль').nth(0).fill(shortUserPassword);
-    await page.getByPlaceholder('Повторіть пароль').fill(shortUserPassword);
-    await page.getByRole('button', {name:"Зареєструватись"}).click()
-    var errorText = await page.locator('#usernameForm .error').innerText()
-    var errorText = errorText.trim()
-    await expect(page.locator('#usernameForm').locator('.error')).toHaveText("* password2 * Пароль надто короткий. Він повинен містити як мінімум 8 символів")
+test('short password registration', async ({page})=>{
+    const registrationPage = new RegistrationPage(page);
+    const validUserName = faker.person.firstName();
+    const shortUserPassword = 'par0l';
+
+    await registrationPage.register(validUserName,shortUserPassword,shortUserPassword);
+    await expect(await registrationPage.getFinalFormErrorText()).toHaveText("* password2 * Пароль надто короткий. Він повинен містити як мінімум 8 символів");
+
 })
 
-test('password is too known', async ({page})=>{
-    const validUserName = faker.person.firstName()
-    const knownPassword = 'abcdefghij'
-    await page.getByPlaceholder('Ім\'я користувача').fill(validUserName);
-    await page.getByPlaceholder('Пароль').nth(0).fill(knownPassword);
-    await page.getByPlaceholder('Повторіть пароль').fill(knownPassword);
-    await page.getByRole('button', {name:"Зареєструватись"}).click()
-    var errorText = await page.locator('#usernameForm .error').innerText()
-    var errorText = errorText.trim()
-    await expect(page.locator('#usernameForm').locator('.error')).toHaveText("* password2 * Пароль надто відомий.")
+test('too known password registration', async({page})=>{
+    const registrationPage = new RegistrationPage(page);
+    const validUserName = faker.person.firstName();
+    const knownPassword = 'abcdefghij';
+
+    await registrationPage.register(validUserName, knownPassword, knownPassword);
+    await expect(await registrationPage.getFinalFormErrorText()).toHaveText("* password2 * Пароль надто відомий.");
 })
 
-test('password integers only', async ({page})=>{
-    const validUserName = faker.person.firstName()
-    const shortUserPassword = '002233114477'
-    await page.getByPlaceholder('Ім\'я користувача').fill(validUserName);
-    await page.getByPlaceholder('Пароль').nth(0).fill(shortUserPassword);
-    await page.getByPlaceholder('Повторіть пароль').fill(shortUserPassword);
-    await page.getByRole('button', {name:"Зареєструватись"}).click()
-    var errorText = await page.locator('#usernameForm .error').innerText()
-    var errorText = errorText.trim()
-    await expect(page.locator('#usernameForm').locator('.error')).toHaveText("* password2 * Цей пароль повністю складається із цифр.")
+test('password from integers only registration', async ({page})=>{
+    const registrationPage = new RegistrationPage(page);
+    const validUserName = faker.person.firstName();
+    const integersPassword = "00223344551177";
+
+    await registrationPage.register(validUserName, integersPassword, integersPassword);
+    await expect(await registrationPage.getFinalFormErrorText()).toHaveText("* password2 * Цей пароль повністю складається із цифр.")
 })
 
-test('passwords do not match', async ({page})=>{
-    const validUserName = faker.person.firstName()
-    const shortUserPassword = '002233114477'
-    await page.getByPlaceholder('Ім\'я користувача').fill(validUserName);
-    await page.getByPlaceholder('Пароль').nth(0).fill(shortUserPassword);
-    await page.getByPlaceholder('Повторіть пароль').fill(shortUserPassword);
-    await page.getByRole('button', {name:"Зареєструватись"}).click()
-    var errorText = await page.locator('#usernameForm .error').innerText()
-    var errorText = errorText.trim()
-    await expect(page.locator('#usernameForm').locator('.error')).toHaveText("* password2 * Паролі не збігаються")
+test('passwords do not match registration', async({page})=>{
+    const registrationPage = new RegistrationPage(page);
+    const validUserName = faker.person.firstName();
+    const validPassword1 = faker.internet.password();
+    const validPassword2 = faker.internet.password();
+
+    await registrationPage.register(validUserName, validPassword1, validPassword2);
+    await expect(await registrationPage.getFinalFormErrorText()).toHaveText("* password2 * Паролі не збігаються");
 })
 
-test('Register with an existing username', async({page})=>{
-    const validUserPassword = faker.internet.password()
-    await page.getByPlaceholder('Ім\'я користувача').fill('Tester');
-    await page.getByPlaceholder('Пароль').nth(0).fill(validUserPassword);
-    await page.getByPlaceholder('Повторіть пароль').fill(validUserPassword);
-
-    var errorNameText = await page.locator('#usernameError').innerText()
-    var errorNameText = errorNameText.trim()
-    await expect(page.locator('#usernameError')).toHaveText("Це ім'я вже зайнято.")
-
-    await page.getByRole('button', {name:"Зареєструватись"}).click()
-    
-    var errorText = await page.locator('#usernameForm .error').innerText()
-    var errorText = errorText.trim()
-    await expect(page.locator('#usernameForm').locator('.error')).toHaveText("* username * Користувач з таким ім'ям вже існує.")
+test('existing username registration', async({page})=>{
+    const registrationPage = new RegistrationPage(page);
+    const validPassword = faker.internet.password();
+    await registrationPage.register("Tester", validPassword, validPassword);
+    await expect(await registrationPage.getFinalFormErrorText()).toHaveText("* username * Користувач з таким ім'ям вже існує.");
 })
-
